Authenticated funcionando

+++ src/controller/tweetController.js
@@ -7,6 +7,10 @@ exports.ADD_TWEET = function (req, res, arrayC) {
     var tweet = new Tweet()
     arrayC.shift()
     var textTweet = arrayC.join(' ')
+
+    if (tweetId != req.user.sub) {
+        return res.status(500).send({ message: 'No posee los permisos para añadir un Tweet' })
+    }
     if (textTweet) {
         tweet.text = textTweet
         tweet.save((err, tweetSave) => {
@@ -24,6 +28,9 @@ exports.ADD_TWEET = function (req, res, arrayC) {
 
 exports.DELETE_TWEET = function (req, res, arrayC) {
 
+    if (tweetId != req.user.sub) {
+        return res.status(500).send({ message: 'No posee los permisos para Eliminar el Tweet' })
+    }
     const tweetId = arrayC[1]
     Tweet.findByIdAndDelete(tweetId, (err, tweetDeleted) => {
         if (err) return res.status(500).send({ message: 'Error en la peticion' })
@@ -38,10 +45,10 @@ exports.EDIT_TWEET = function (req, res, arrayC) {
     arrayC.shift()
     arrayC.shift()
     var newTweet = arrayC.join(" ")
-    console.log(tweetId);
-    
-    console.log(newTweet);
-    
+
+    if (tweetId != req.user.sub) {
+        return res.status(500).send({ message: 'No posee los permisos para Editar el Tweet' })
+    }
     Tweet.findByIdAndUpdate(tweetId, newTweet, { new: true }, (err, tweetUpdated) => {
         if (err) return res.status(500).send({ message: 'Error en la peticion' })
         if (!tweetUpdated) return res.status(404).send({ message: 'No se ha podido editar el Tweet' })
@@ -53,4 +60,3 @@ exports.EDIT_TWEET = function (req, res, arrayC) {
 exports.VIEW_TWEETS = function (req, res, arrayC) {
 
 }
-
